requirements.txt
langchain==0.1.0
chromadb==0.4.22
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1 
langchain-community
sentence-transformers